Fixed audio stream system
Reorganized the ogg file.
Added comments to clarify how setup of stream works.
Moved format detection to top of setup function instead of repeating it
for each buffer.

Added a flag to the stream struct to not rewind it the first time it is
played, as it already has data about the start of the audio file.